pyproject.toml
[project]
name = "daily-digest-prompt"
version = "0.1.0"
description = "Automated daily digest system with Claude Code"
requires-python = ">=3.9"
# 核心依賴使用 == 鎖定（ADR-010：防止惡意版本注入；版本與 uv.lock 對齊）
# fpdf2 維持 >= 以相容 requires-python >=3.9（fpdf2 2.8.7 需 3.10+）
dependencies = [
    "requests==2.32.5",
    "schedule==1.2.2",
    "pyyaml==6.0.3",
    "edge-tts>=6.1.0",
    "fpdf2>=2.8.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "pip-audit>=2.0.0",
]
gmail = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.80.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --cov=hooks --cov=skills --cov-report=term-missing --cov-fail-under=44"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]
